pocket-ic must be in dev dependencies
