+++ e2e-tests/Cargo.toml
@@ -10,15 +10,15 @@ publish = false
 
 [dependencies]
 candid.workspace = true
-cargo_metadata = "0.18"
 escargot = { version = "0.5.7", features = ["print"] }
+futures.workspace = true
 ic-cdk.workspace = true
 ic-cdk-timers.workspace = true
 lazy_static = "1.4.0"
 serde_bytes.workspace = true
-futures.workspace = true
-sha2.workspace = true
-pocket-ic = { git = "https://github.com/dfinity/ic", tag = "release-2024-11-28_03-15-base" }
 
 [dev-dependencies]
+cargo_metadata = "0.19"
 hex.workspace = true
+pocket-ic = { git = "https://github.com/dfinity/ic", tag = "release-2024-11-28_03-15-base" }
+sha2.workspace = true
