+++ e2e-tests/tests/api_call.rs
@@ -1,7 +1,9 @@
 use candid::{Encode, Principal};
-use ic_cdk_e2e_tests::{cargo_build_canister,pocket_ic};
 use pocket_ic::common::rest::RawEffectivePrincipal;
-use pocket_ic::{call_candid, query_candid, PocketIc, WasmResult};
+use pocket_ic::{call_candid, query_candid,  WasmResult};
+
+mod test_utilities;
+use test_utilities::{cargo_build_canister,pocket_ic};
 
 #[test]
 fn call_apis() {
@@ -41,7 +43,7 @@ fn call_apis() {
 
 #[test]
 fn cycles_burn() {
-    let pic = PocketIc::new();
+    let pic = pocket_ic();
     let wasm = cargo_build_canister("api_call");
     let canister_id = pic.create_canister();
     pic.add_cycles(canister_id, 2_000_000_000_000);
