+++ e2e-tests/tests/call_struct.rs
@@ -1,11 +1,13 @@
 use candid::Principal;
-use ic_cdk_e2e_tests::cargo_build_canister;
+use pocket_ic::call_candid;
 use pocket_ic::common::rest::RawEffectivePrincipal;
-use pocket_ic::{call_candid, PocketIc};
+
+mod test_utilities;
+use test_utilities::{cargo_build_canister, pocket_ic};
 
 #[test]
 fn call_struct() {
-    let pic = PocketIc::new();
+    let pic = pocket_ic();
     let wasm = cargo_build_canister("call_struct");
     let canister_id = pic.create_canister();
     pic.add_cycles(canister_id, 100_000_000_000_000);
