+++ e2e-tests/tests/simple_kv_store.rs
@@ -1,14 +1,16 @@
-use ic_cdk_e2e_tests::cargo_build_canister;
 use pocket_ic::common::rest::RawEffectivePrincipal;
-use pocket_ic::{call_candid, query_candid, PocketIc};
+use pocket_ic::{call_candid, query_candid};
 use serde_bytes::ByteBuf;
+
+mod test_utilities;
+use test_utilities::{cargo_build_canister, pocket_ic};
 
 /// Checks that a canister that uses [`ic_cdk::storage::stable_save`]
 /// and [`ic_cdk::storage::stable_restore`] functions can keep its data
 /// across upgrades.
 #[test]
 fn test_storage_roundtrip() {
-    let pic = PocketIc::new();
+    let pic = pocket_ic();
     let wasm = cargo_build_canister("simple_kv_store");
     let canister_id = pic.create_canister();
     pic.add_cycles(canister_id, 2_000_000_000_000);
